Investigating difference in k_Advection from regional to continental air.
TotalArea input in ScaleSheet.xlsx for Regional does not match that in Excel; rounding error. Should be 2.2957E11 instead of 2.30E11

k.aR.aC in excel does match R k_Adevction for regional if area in excel is changed to match R input

So no fault of the formules in R, but due to different inputs.

As the implementation of the total area differs between R and Excel. Due to the nesting in R, I am unsure how exactly to address this; should the TotalAreas of continental and moderate zones also be changed when changing that of Regional?